frontend: extract address or txid into component
Create TxAddress and TxAddressOrTxID component that can be used to
display transaciton addresses or the txid in the details and the
transaction itself. TypeScript also enforces that either addresses OR
txid is passed to TxAddressOrTxID but not both. This makes the
transaction.tsx component easier to read and more concise.