backend: add isChange to utxos response
Add a boolean field to the response of the utxos enpoint (the getUTXOS
handler) that tells us whether the utxo is to a change address of our
wallet or not. Possible false negative in case the user used another
wallet and derived some change address behind the current change gap
limit.